Reindex algorithm - move reindex struct and all related operations to new file
